Update GitHub Actions workflow for website deployment (#13)
The GitHub Actions workflow for deploying the static website to GitHub
Pages has been updated. The unnecessary artifacts job has been removed,
and the build job has been renamed to `build` and updated with the
necessary steps for building the website. The deployment job has also
been updated to have a dependency on the `build` job.