Refactor project structure and update welcome message (#30)
This pull request includes several changes to improve the project
structure and user experience. The commits refactor the project
structure by moving the Website (Next.js project) to the Client folder,
update the tsconfig.json file to remove outdated paths and update type
definitions for Next.js, and update the build and artifact paths in the
GitHub Actions workflow. Additionally, the welcome message in the
page.tsx file is updated to include the application name for clarity.
These changes enhance the organization of the codebase and improve the
user experience by providing a more informative welcome message.