Format configuration files and add Storybook support (#35)
This pull request includes various commits that format configuration
files for consistency and readability. It also adds support for
Storybook configuration in the client app, including the initialization
of Storybook configuration files and the addition of necessary
dependencies. These changes enhance the development experience with
stories and improve UI component testing and documentation.